startserver/stopserver stuff, luasnip autocompletion, clangd config
The clangd LSP is now configured to never automatically import headers,
which is more convenient for me. Added some tools to make it easy for
other applications (in this case, Sublime Merge) to open a file in
Neovim. Added a command to make it easier to edit the configuration for
Neovim (`ConfigExplore` takes you right to the config directory). Also
added autocompletion support for snippets added manually with LuaSnip.